+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 #### :nail_care: Polish
 
 - Avoid running `rescript-schema-ppx` and `sury-ppx` on source files without an `@schema` annotation. https://github.com/rescript-lang/rescript/pull/8662
+- Format JSX expression children and standalone comments with braces to prepare for future literal text support. https://github.com/rescript-lang/rescript/pull/8666
 
 #### :house: Internal